backend/requirements.txt
# FastAPI & ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# MikroTik clients
routeros-api==0.17.0
paramiko==3.4.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
pydantic>=2.10
pydantic-settings>=2.7

# HTTP client
httpx==0.26.0

# Reports
pandas>=2.2
openpyxl>=3.1

# Monitoring & Logging
structlog==24.1.0

# Rate limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
